requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-cleanup==5.2.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-random-queryset==0.1.3
django-tinymce==2.8.0
djangorestframework==3.12.4
idna==2.10
more-itertools==8.10.0
Pillow==9.4.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
PyMySQL==1.0.2
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.2
xmltodict==0.12.0
